+++ src/Utils/HandleModel.php
@@ -2,7 +2,7 @@
 
 namespace Sfneal\CrudModelActions\Utils;
 
-use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Model as EloquentModel;
 
 // todo: refactor to interface
 trait HandleModel
@@ -10,7 +10,7 @@ trait HandleModel
     /**
      * Save or update the Model.
      *
-     * @return Model
+     * @return EloquentModel
      */
-    abstract protected function handle(): Model;
+    abstract protected function handle(): EloquentModel;
 }